Refactor: Rename Tail into DirectoryStats and tail.rs to stats.rs
