+++ src/config/inspect.rs
@@ -1,20 +1,20 @@
 use crate::error::simple::TResult;
-use crate::report::tail::Tail;
+use crate::report::stats::DirectoryStats;
 
 use std::fs;
 use std::fs::DirEntry;
 use std::path::PathBuf;
 
-pub type FnReadDir = fn(PathBuf, &mut Tail) -> TResult<Vec<DirEntry>>;
+pub type FnReadDir = fn(PathBuf, &mut DirectoryStats) -> TResult<Vec<DirEntry>>;
 
 #[allow(unused_variables)]
-pub fn read_all_entries(path: PathBuf, tail: &mut Tail) -> TResult<Vec<DirEntry>> {
+pub fn read_all_entries(path: PathBuf, tail: &mut DirectoryStats) -> TResult<Vec<DirEntry>> {
     let entries = fs::read_dir(path)?.collect::<Result<Vec<DirEntry>, std::io::Error>>()?;
 
     Ok(entries)
 }
 
-pub fn read_visible_entries(path: PathBuf, tail: &mut Tail) -> TResult<Vec<DirEntry>> {
+pub fn read_visible_entries(path: PathBuf, tail: &mut DirectoryStats) -> TResult<Vec<DirEntry>> {
     let entries = fs::read_dir(path)?
         .filter_map(|entry_result| {
             entry_result.ok().and_then(|entry| {
@@ -30,7 +30,7 @@ pub fn read_visible_entries(path: PathBuf, tail: &mut Tail) -> TResult<Vec<DirEn
     Ok(entries)
 }
 
-pub fn read_visible_folders(path: PathBuf, tail: &mut Tail) -> TResult<Vec<DirEntry>> {
+pub fn read_visible_folders(path: PathBuf, tail: &mut DirectoryStats) -> TResult<Vec<DirEntry>> {
     let entries = fs::read_dir(path)?
         .filter_map(|entry_result| {
             entry_result.ok().and_then(|entry| {
