+++ src/config/registry.rs
@@ -18,7 +18,7 @@ use crate::render::entree::dirr::FnOutDir;
 use crate::render::entree::filee::FnOutFile;
 use crate::render::entree::headd::FnOutHead;
 use crate::render::entree::symlinked::FnOutSymlink;
-use crate::report::tail::Tail;
+use crate::report::stats::DirectoryStats;
 
 use std::fs::DirEntry;
 use std::io::StdoutLock;
@@ -44,7 +44,7 @@ pub struct Registry<'a> {
 }
 
 impl<'a> Registry<'a> {
-    pub fn inspt_dents(&self, path: PathBuf, tail: &mut Tail) -> TResult<Vec<DirEntry>> {
+    pub fn inspt_dents(&self, path: PathBuf, tail: &mut DirectoryStats) -> TResult<Vec<DirEntry>> {
         (self.read)(path, tail)
     }
