+++ src/report/mod.rs
@@ -1,4 +1,4 @@
 pub mod depth;
 pub mod size;
-pub mod tail;
+pub mod stats;
 pub mod widest;
